requirements.txt
Django~=4.2.16
djangorestframework~=3.15.2
django-filter~=24.3
requests~=2.32.3
orjson~=3.10.7
psycopg2-binary~=2.9.9
drf-yasg~=1.21.7
factory-boy~=3.3.1